requirements.txt
# requirements.txt for VPS API Confirmation Server
fastapi
uvicorn
sqlalchemy
pymysql
pika
alembic
pydantic
python-dotenv